+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Changelog
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+2.0.0
+*****
+* Simulator deploys to ADR-9 TRLs by default. Deployment to the old pre-ADR-9 TRLs can be triggered by setting TMCSIM_USE_OLD_TRLS environment variable at deployment time.
+* *BREAKING* CentralNode device uses correct Tango nomenclature, with device property now named 'domain' instead of 'base_uri'.
+* *BREAKING* TMCSimTestHarness now requires a domain to be set and does not default to TMC-Mid.
+
 1.0.0
 *****
 * Renamed python package to `ska_oso_tmcsim` from `ska_oso_tmc_integration_tests.tmcsim`